requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-telegram-bot==20.6
httpx==0.25.2
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4
feedparser
pydantic>=1.7.4,<3.0.0
ujson>=4.0.1
orjson>=3.2.1
email_validator>=2.0.0
starlette>=0.37.2,<0.38.0
fastapi-cli>=0.0.2
python-multipart>=0.0.7
jinja2>=2.11.2
typing-extensions>=4.8.0
uvloop>=0.14.0
watchfiles>=0.13
httptools>=0.5.0
websockets>=10.4
pydantic-core==2.41.5
annotated-types>=0.6.0
typing-inspection>=0.4.2
rich>=13.7.1
mdurl~=0.1
markdown-it-py>=2.2.0
sniffio>=1.3.0
anyio>=4.11.0
sgmllib3k



